发布版本9.0.0

+++ Agora-Video-With-FaceUnity-iOS/BeautifyExample/FaceUnity/Demo/Model/FUBeautyShapeModel.h
@@ -1,11 +1,12 @@
 //
 //  FUBeautyShapeModel.h
-//  FUDemo
+//  FUBeautyComponent
 //
-//  Created by 项林平 on 2021/7/19.
+//  Created by 项林平 on 2022/7/8.
 //
 
 #import <Foundation/Foundation.h>
+#import <FURenderKit/FURenderKit.h>
 #import "FUDefines.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -16,9 +17,9 @@ NS_ASSUME_NONNULL_BEGIN
 /// 美型类型
 @property (nonatomic, assign) FUBeautyShape type;
 /// 当前值
-@property (nonatomic) double currentValue;
+@property (nonatomic) float currentValue;
 /// 默认值
-@property (nonatomic) double defaultValue;
+@property (nonatomic) float defaultValue;
 /// 默认值是否中位数
 @property (nonatomic, assign) BOOL defaultValueInMiddle;
 /// 设备性能等级要求
