+++ Agora-Video-With-FaceUnity-iOS/BeautifyExample/FaceUnity/Demo/Model/FUBeautySkinModel.h
@@ -1,11 +1,12 @@
 //
 //  FUBeautySkinModel.h
-//  FUDemo
+//  FUBeautyComponent
 //
-//  Created by 项林平 on 2021/7/19.
+//  Created by 项林平 on 2022/7/8.
 //
 
 #import <Foundation/Foundation.h>
+#import <FURenderKit/FURenderKit.h>
 #import "FUDefines.h"
 
 NS_ASSUME_NONNULL_BEGIN
@@ -16,9 +17,9 @@ NS_ASSUME_NONNULL_BEGIN
 /// 美肤类型
 @property (nonatomic, assign) FUBeautySkin type;
 /// 当前值
-@property (nonatomic, assign) double currentValue;
+@property (nonatomic, assign) float currentValue;
 /// 默认值
-@property (nonatomic, assign) double defaultValue;
+@property (nonatomic, assign) float defaultValue;
 /// 默认值是否中位数
 @property (nonatomic, assign) BOOL defaultValueInMiddle;
 /// 实际值对应0.0-1.0的倍率
